+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "fetch-ax",
-  "version": "1.0.3",
+  "version": "1.0.5",
   "description": "A modern HTTP client that extends the Fetch API, providing Axios-like syntax and full compatibility with Next.js App Router.",
   "types": "./lib/cjs/types/index.d.ts",
   "main": "./lib/cjs/index.js",
@@ -18,7 +18,14 @@
     "type": "git",
     "url": "git+https://github.com/Myongji-Graduate/fetch-ax.git"
   },
-  "keywords": [],
+  "keywords": [
+    "fetch",
+    "axios",
+    "http",
+    "client",
+    "fetch-ax",
+    "Next.js"
+  ],
   "author": "gahyuun",
   "license": "ISC",
   "dependencies": {
