--- src/preset-options.ts
@@ -1,9 +1,0 @@
-import { FetchAXDefaultOptions } from '.';
-
-export const presetOptions: FetchAXDefaultOptions = {
-  headers: new Headers([['Content-Type', 'application/json']]),
-
-  throwError: true,
-
-  responseType: 'json',
-};
